+++ src-tauri/src/login.rs
@@ -1,0 +1,96 @@
+use std::borrow::Cow;
+
+use reqwest::Url;
+use tauri::Window;
+use tauri_plugin_oauth::OauthConfig;
+
+use crate::{db, utils::open_path};
+
+const OAUTH_CLIENT_ID: &str = "github.com/fitztrev/lichess-tauri";
+
+#[allow(dead_code)]
+#[derive(serde::Deserialize, Debug)]
+struct AccessTokenResponse {
+    token_type: String,
+    access_token: String,
+    expires_in: i32,
+}
+
+#[derive(serde::Deserialize, Debug)]
+struct LichessAccount {
+    username: String,
+}
+
+pub fn start_oauth_flow(window: Window) {
+    let (code_challenge, code_verify) = oauth2::PkceCodeChallenge::new_random_sha256();
+
+    let port = tauri_plugin_oauth::start_with_config(
+        OauthConfig {
+            ports: None,
+            response: Some(Cow::Borrowed(include_str!("../public/oauth_response.html"))),
+        },
+        move |url| {
+            println!("returning_from_lichess: {}", url);
+
+            let url = Url::parse(&url).unwrap();
+            println!("url: {:?}", url);
+
+            let code = url.query_pairs().find(|(key, _)| key == "code").unwrap().1;
+
+            println!("code_verify: {}", code_verify.secret());
+
+            let lichess_host = db::get_setting("lichess_host").unwrap();
+
+            let body = reqwest::blocking::Client::new()
+                .post(format!("{}/api/token", lichess_host))
+                .form(&[
+                    ("grant_type", "authorization_code"),
+                    ("client_id", OAUTH_CLIENT_ID),
+                    ("code", code.to_string().as_str()),
+                    (
+                        "redirect_uri",
+                        format!("http://localhost:{}/", url.port().unwrap()).as_str(),
+                    ),
+                    ("code_verifier", code_verify.secret()),
+                ])
+                .send()
+                .unwrap()
+                .json::<AccessTokenResponse>()
+                .unwrap();
+
+            println!("body: {:?}", body);
+
+            // update db with access token
+            db::update_setting("lichess_token", &body.access_token);
+
+            let me = reqwest::blocking::Client::new()
+                .get(format!("{}/api/account", lichess_host))
+                .bearer_auth(&body.access_token)
+                .send()
+                .unwrap()
+                .json::<LichessAccount>()
+                .unwrap();
+
+            println!("me: {:?}", me);
+
+            db::update_setting("lichess_username", &me.username);
+
+            window.emit("refresh_settings_from_database", ()).unwrap();
+        },
+    )
+    .unwrap();
+
+    println!("Local server started on port: {}", port);
+    let redirect_url = format!("http://localhost:{}/", port);
+
+    let lichess_host = db::get_setting("lichess_host").unwrap();
+    let url = format!(
+        "{}/oauth?response_type=code&client_id={}&redirect_uri={}&code_challenge_method=S256&code_challenge={}&scope=engine:read%20engine:write",
+        lichess_host,
+        OAUTH_CLIENT_ID,
+        redirect_url,
+        code_challenge.as_str()
+    );
+
+    open_path(url);
+}
