+++ src-tauri/src/utils.rs
@@ -13,8 +13,9 @@ pub fn get_app_data_dir() -> path::PathBuf {
     app_data_dir
 }
 
-// https://github.com/fitztrev/lichess-tauri/issues/26
-// implementing solution from: https://github.com/tauri-apps/tauri/issues/6172#issuecomment-1557017115
+/// Opens a URL or file path in the OS default application
+/// https://github.com/fitztrev/lichess-tauri/issues/26
+/// implementing solution from: https://github.com/tauri-apps/tauri/issues/6172#issuecomment-1557017115
 pub fn open_path(path: String) {
     // `open` required to run in separate thread, to avoid blocking on some
     // platforms (eg Fedora38 blocks)
